Created a resusable hamburger menu object
