+++ team_b/yappy/lib/restaurant.dart
@@ -11,22 +11,5 @@ class RestaurantApp extends StatelessWidget {
 }
 
 class RestaurantPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.black,
-      appBar: AppBar(
-        backgroundColor: Colors.black,
-        leading: Builder(
-          builder: (context) {
-
-            //I would like to figure out how to change the drawer size
-            return IconButton(
-              icon: const Icon(Icons.menu, color: Colors.white),
-              onPressed: () {
-              Scaffold.of(context).openDrawer();
-              },
-            );
-          },        
-        ),
+    
 }
